git-svn-id: svn://192.168.1.10/public/CoffeeMud@11210 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/core/threads/StdTickClient.java
@@ -28,7 +28,7 @@ public class StdTickClient implements TickClient
 	public volatile long   lastStart=0;
 	public volatile long   lastStop=0;
 	public volatile long   milliTotal=0;
-	public volatile long   tickTotal=0;
+	public volatile int	   tickTotal=0;
 	public volatile String status = null;
 
 	public StdTickClient(Tickable newClientObject, int newTickDown, int newTickID)
